backend/requirements.txt
# Flask and API dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-Session>=0.5.0
Werkzeug>=2.3.0

# MongoDB
mongoengine>=0.27.0
pymongo>=4.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# Blockchain dependencies (from blockchain module)
web3>=6.0.0
eth-account>=0.11.0

# HTTP requests (for ML service integration)
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.2

# JSON handling
jsonschema>=4.19.0

# Logging
colorlog>=6.7.0

# DNS resolver for MongoDB SRV records
dnspython>=2.4.0

# Twilio WhatsApp integration
twilio>=8.0.0

# Google Dialogflow integration
google-cloud-dialogflow>=2.0.0
# Email service
sendgrid>=6.9.0

# QR Code generation
qrcode[pil]>=7.4.2
Pillow>=10.0.0